requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
gunicorn==21.2.0
Werkzeug==2.2.3

tensorflow-cpu==2.20.0
numpy==1.26.4
nltk==3.9.1
pickle-mixin==1.0.2
